backend/requirements.txt
# --- Core Framework ---
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.3.4

# --- AI & Orchestration ---
langgraph==0.0.65
langchain==0.2.1
google-generativeai==0.7.2
weaviate-client[agents]==4.9.0

# --- Observability (OpenTelemetry) ---
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-exporter-otlp==1.25.0

# --- Utilities ---
pypdf==4.2.0
redis==5.0.4
httpx==0.27.0
python-multipart==0.0.9